3.3 agent flow refactoring
