+++ features/ai/components/flow/StepExecutor.tsx
@@ -1,0 +1,131 @@
+'use client';
+import React, { useEffect } from 'react';
+import { useAgentFlow } from '@/providers/AgentFlowProvider';
+import { selectBestDesignConcept } from '@/lib/specSelection';
+import { useUserGuid } from '@/providers/UserGuidProvider';
+
+interface StepExecutorProps {
+  brief: string;
+  setBrief: (v: string) => void;
+}
+
+export default function StepExecutor({ brief, setBrief }: StepExecutorProps) {
+  const {
+    steps,
+    currentStep,
+    completeStep,
+    abort,
+    aborted,
+    startExecution,
+    designConcepts,
+    setDesignConcepts,
+    evaluationResults,
+    setEvaluationResults,
+    selectedConcept,
+    setSelectedConcept,
+    addError,
+    failedStep
+  } = useAgentFlow();
+
+  const userGuid = useUserGuid();
+
+  const startFlow = async () => {
+    if (currentStep <= 0) {
+      startExecution();
+      try {
+        const res = await fetch('/api/agent/generate-design-concepts', {
+          method: 'POST',
+          headers: { 'Content-Type': 'application/json', 'x-user-guid': userGuid },
+          body: JSON.stringify({ brief })
+        });
+        const data = await res.json();
+        if (Array.isArray(data.concepts)) {
+          setDesignConcepts(data.concepts);
+          completeStep(0);
+        } else {
+          addError('Failed to generate design concepts', 0);
+        }
+      } catch (err) {
+        const message = err instanceof Error ? err.message : 'Unknown error';
+        addError(message, 0);
+      }
+    }
+  };
+
+  const nextStep = async () => {
+    if (currentStep < steps.length) {
+      if (currentStep === 1 && designConcepts.length > 0) {
+        try {
+          const res = await fetch('/api/agent/evaluate-designs', {
+            method: 'POST',
+            headers: { 'Content-Type': 'application/json', 'x-user-guid': userGuid },
+            body: JSON.stringify({ concepts: designConcepts })
+          });
+          const data = await res.json();
+          if (Array.isArray(data.evaluations)) {
+            setEvaluationResults(data.evaluations);
+            setSelectedConcept(selectBestDesignConcept(data.evaluations));
+            completeStep(currentStep);
+          } else {
+            addError('Failed to evaluate designs', currentStep);
+          }
+        } catch (err) {
+          const message = err instanceof Error ? err.message : 'Unknown error';
+          addError(message, currentStep);
+        }
+        return;
+      }
+      completeStep(currentStep);
+    }
+  };
+
+  useEffect(() => {
+    if (currentStep > 0 && currentStep < 3 && !aborted && failedStep === null) {
+      nextStep();
+    }
+  }, [currentStep, aborted, failedStep]);
+
+  return (
+    <>
+      {currentStep <= 0 && !aborted && (
+        <div className="bg-white rounded-2xl shadow-lg border border-gray-100 p-8 mb-8">
+          <div className="max-w-2xl mx-auto">
+            <h2 className="text-2xl font-semibold text-gray-900 mb-2">Start Your Journey</h2>
+            <p className="text-gray-600 mb-6">Describe your vision and let AI bring it to life</p>
+            <div className="space-y-4">
+              <label className="block">
+                <span className="text-sm font-medium text-gray-700 mb-2 block">Creative Brief</span>
+                <textarea
+                  className="w-full p-4 border border-gray-200 rounded-xl focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all duration-200 resize-none bg-gray-50 focus:bg-white"
+                  rows={5}
+                  value={brief}
+                  onChange={(e) => setBrief(e.target.value)}
+                  placeholder="Describe the UI component you want to create. Be specific about functionality, style, and data visualization needs..."
+                />
+              </label>
+              <button
+                onClick={startFlow}
+                disabled={!brief.trim()}
+                className="w-full bg-gradient-to-r from-blue-600 to-purple-600 text-white py-4 px-6 rounded-xl font-semibold text-lg shadow-lg hover:shadow-xl transform hover:-translate-y-0.5 transition-all duration-200 disabled:opacity-50 disabled:cursor-not-allowed disabled:transform-none disabled:shadow-lg"
+              >
+                {brief.trim() ? '🚀 Start AI Agent Flow' : 'Enter your creative brief to begin'}
+              </button>
+            </div>
+          </div>
+        </div>
+      )}
+      {!aborted && currentStep >= 0 && currentStep < steps.length && (
+        <div className="bg-white rounded-2xl shadow-lg border border-gray-100 p-8 mb-8">
+          <div className="flex flex-col sm:flex-row gap-4 justify-center">
+            <button
+              onClick={() => abort()}
+              className="bg-gradient-to-r from-red-500 to-pink-500 text-white py-3 px-8 rounded-xl font-semibold shadow-lg hover:shadow-xl transform hover:-translate-y-0.5 transition-all duration-200"
+            >
+              Abort Flow
+            </button>
+          </div>
+        </div>
+      )}
+    </>
+  );
+}
