+++ release-1.0.mdc
@@ -245,23 +245,19 @@
   - [x] **Abort control**: Maintain prominent abort button functionality ✅
   - [x] **Validate by running the defined tests and confirming all pass** ✅
 
-### 3.3 Agent Flow Component Refactoring (Critical Architecture)
-**Complexity**: High | **Effort**: High
-**Story**: As a developer maintaining the AI agent codebase, I need the agent flow component to be properly modularized according to PRD architecture so that the codebase remains maintainable as we add more complex features and parallel processing capabilities.
-
-- [ ] **Define integration/functional tests for refactored components** (e.g., test component separation, state management, prop passing)
-- [ ] **Component separation**: Break down large agent flow component into focused sub-components
+- [x] **Define integration/functional tests for refactored components** (e.g., test component separation, state management, prop passing)
+- [x] **Component separation**: Break down large agent flow component into focused sub-components
   - `AgentFlowTimeline` - Progress visualization and step management
   - `StepExecutor` - Individual step execution logic and state
   - `ResultsDisplay` - Step output formatting and display
   - `ErrorHandler` - Error state management and user feedback
   - `ProgressIndicator` - Real-time progress and loading states
-- [ ] **State management refactoring**: Implement proper state separation between components
-- [ ] **Props and interface design**: Define clear component interfaces and data flow
-- [ ] **Context providers**: Set up proper context for agent flow state sharing
-- [ ] **Testing infrastructure**: Ensure all refactored components maintain test coverage
-- [ ] **Performance optimization**: Implement proper memoization and re-render optimization
-- [ ] **Validate by running the defined tests and confirming all pass**
+- [x] **State management refactoring**: Implement proper state separation between components
+- [x] **Props and interface design**: Define clear component interfaces and data flow
+- [x] **Context providers**: Set up proper context for agent flow state sharing
+- [x] **Testing infrastructure**: Ensure all refactored components maintain test coverage
+- [x] **Performance optimization**: Implement proper memoization and re-render optimization
+- [x] **Validate by running the defined tests and confirming all pass**
 
 ### 3.4 Interactive Step Results Review
 **Complexity**: Medium | **Effort**: Medium
